LPS-116840 showSidebar no longer needs to be outside the propsTransformer
